Started moving parts into abstractions projects
