+++ src/Server/AspNetClassic.Abstractions/AspNetClassic.Abstractions.csproj
@@ -15,13 +15,13 @@
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
     <DebugType>portable</DebugType>
     <DebugSymbols>true</DebugSymbols>
-    <DefineConstants>TRACE;DEBUG</DefineConstants>
+    <DefineConstants>TRACE;DEBUG;ASPNETCLASSIC</DefineConstants>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
     <DebugType>pdbonly</DebugType>
     <DebugSymbols>true</DebugSymbols>
-    <DefineConstants>TRACE</DefineConstants>
+    <DefineConstants>TRACE;ASPNETCLASSIC</DefineConstants>
   </PropertyGroup>
 
   <ItemGroup>
