+++ src/Server/AspNetClassic/AspNetClassic.csproj
@@ -32,6 +32,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Core\Core\Core.csproj" />
     <ProjectReference Include="..\AspNetClassic.Authorization\AspNetClassic.Authorization.csproj" />
+    <ProjectReference Include="..\AspNetClassic.Abstractions\AspNetClassic.Abstractions.csproj" />
     <ProjectReference Include="..\Server\Server.csproj" />
   </ItemGroup>
