+++ src/Server/AspNetClassic/EnvironmentKeys.cs
@@ -1,11 +1,7 @@
-﻿#if ASPNETCLASSIC
-
-namespace HotChocolate.AspNetClassic
+﻿namespace HotChocolate.AspNetClassic
 {
     internal static class EnvironmentKeys
     {
         public const string ServiceProvider = "hotchocolate.ServiceProvider";
     }
 }
-
-#endif
