+++ src/Server/AspNetCore.Abstractions/AspNetCore.Abstractions.csproj
@@ -25,9 +25,11 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="System.IO.Pipelines" Version="4.5.3" />
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\Core\Core\Core.csproj" />
   </ItemGroup>
 
 </Project>
